refactor(solver): extract foreign-param-shape helpers to sibling module (size ratchet)
Move is_bare_foreign_type_param and is_bare_foreign_type_param_shape from
inference_helpers.rs into a new generic_call/foreign_param_shapes.rs sibling
module, bringing inference_helpers.rs back under its 2065-line ratchet ceiling
(2085 -> 2010). Pure relocation: identical function names, signatures, and
semantics; call sites in inference_helpers.rs and resolve/finalize.rs updated
to the new module path.
